tflite_unit_tests: split into own file
Moves the unit tests to their own file, matching the functionality
exposed by tflite_unit_tests.h.
